move chose_qgt
