+++ netket/jax/_jacobian/__init__.py
@@ -1,1 +1,2 @@
 from .logic import jacobian
+from .default_mode import jacobian_default_mode
