+++ netket/optimizer/qgt/qgt_jacobian_dense.py
@@ -28,7 +28,6 @@ from ..linear_operator import LinearOperator, Uninitialized
 
 from .common import check_valid_vector_type
 from .qgt_jacobian_common import (
-    choose_jacobian_mode,
     sanitize_diag_shift,
     to_shift_offset,
     rescale,
@@ -112,7 +111,7 @@ def QGTJacobianDense(
         pdf = None
 
     if mode is None:
-        mode = choose_jacobian_mode(
+        mode = nkjax.jacobian_default_mode(
             vstate._apply_fun,
             vstate.parameters,
             vstate.model_state,
